+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Display user full name in logs in addition to username.
 
+### Fixed
+
+- The issue where running the `/getconfig` command would overwrite the OpenAI API key.
+
 ## [0.4.0] - 2025-03-22
 
 ### Added